Add CSS post-processing with Lightning CSS
This commit adds `.css` as a custom template extension to eleventy to be
bundled by lightningcss.

Per Sass convention, css files that start with an underscore are treated
as partials and therefore not compiled into the output folder. I also
excluded css files inside a directory starting with an underscore from
the build.