Fix bug in Firefox
This commit refactors the previous implementation for getting the form
fields, which works in Chrome but not in Firefox. The issue is caused by
the different browser implementations of `Object.keys()` and/or the
`HTMLFormControlsCollection`:

- In Chrome, the array returned by `Object.keys(form.elements)` includes
  the name of inputs that have a `name` attribute, since they are
  treated as enumerable properties.

- In Firefox, those same properties are treated as **non-enumerable** so
  they're not included in the returned array.

Using `Object.getOwnPropertyNames()` was my first option since it
includes non-enumerable properties. But I only want the "non-enumerable
properties", so I resorted to filtering the form elements to only the
ones that have a `name` attribute.